# swagger/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API documentation",
        "title": "Go Boilerplate API",
        "contact": {
            "name": "Abdan Hafidz",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/events": {
            "post": {
                "description": "Create a new event with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Create Event",
                "parameters": [
                    {
                        "description": "Create Event Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-dto_EventDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/events/{id}": {
            "put": {
                "description": "Update an existing event with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Update Event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Event Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-models_Events"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing event by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Delete Event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-map_string_bool"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/events": {
            "get": {
                "description": "Retrieve a paginated list of events with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "List Events",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term for event titles",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Field to sort by (e.g., 'date', 'title')",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order ('asc' or 'desc')",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by registration status",
                        "name": "registerStatus",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by event status (upcoming, ongoing, ended)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-array_models_Events"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/events/register-event": {
            "post": {
                "description": "Register the authenticated user for an event using an event code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Join Event",
                "parameters": [
                    {
                        "description": "Join Event Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.JoinEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-dto_EventDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-models_EventPaymentTransaction"
                        }
                    }
                }
            }
        },
        "/api/v1/events/{event_slug}": {
            "get": {
                "description": "Retrieve detailed information about a specific event using its slug",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Get Event Detail by Slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Slug",
                        "name": "event_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-dto_EventDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateEventRequest": {
            "type": "object",
            "required": [
                "end_event",
                "event_code",
                "img_banner",
                "overview",
                "start_event",
                "title"
            ],
            "properties": {
                "end_event": {
                    "type": "string"
                },
                "event_code": {
                    "type": "string"
                },
                "img_banner": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "overview": {
                    "type": "string"
                },
                "start_event": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.EventDetailResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Events"
                },
                "event_payment": {
                    "$ref": "#/definitions/models.EventPaymentTransaction"
                },
                "register_status": {
                    "type": "integer"
                }
            }
        },
...
        "dto.JoinEventRequest": {
            "type": "object",
            "required": [
                "event_code"
            ],
            "properties": {
                "event_code": {
                    "type": "string"
                }
            }
        },
...
        "dto.SuccessResponse-array_models_Events": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Events"
                    }
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.SuccessResponse-dto_EventDetailResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.EventDetailResponse"
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.SuccessResponse-map_string_bool": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/map_string_bool"
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.SuccessResponse-models_EventPaymentTransaction": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.EventPaymentTransaction"
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.SuccessResponse-models_Events": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Events"
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateEventRequest": {
            "type": "object",
            "properties": {
                "end_event": {
                    "type": "string"
                },
                "img_banner": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "overview": {
                    "type": "string"
                },
                "start_event": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "map_string_bool": {
            "type": "object",
            "additionalProperties": {
                "type": "boolean"
            }
        },
...
        "models.EventPaymentTransaction": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "event_id": {
                    "type": "string"
                },
                "expired_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_id": {
                    "type": "string"
                },
                "invoice_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transaction_at": {
                    "type": "string"
                },
                "xendit_transaction_id": {
                    "type": "string"
                }
            }
        },
        "models.Events": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "end_event": {
                    "type": "string"
                },
                "event_code": {
                    "type": "string"
                },
                "id_event": {
                    "type": "string"
                },
                "img_banner": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "overview": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "register_status": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "start_event": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}